Make map `kind` an enum; Fix `modes` type
The `kind` has a specific set of things that it can be. It didn't need
to be a `String`.

Fix `DKMapGroup.modes` to ensure that values are map hashes.

Remove `trigger` from `Map` because it didn't make sense to include it
in the `HashMap` value if it's the hash key.