run_key_action(): Include action string for `MapKind::Map`
Couldn't figure out how to make my `Action` string a char slice. At
first I changed it to a `CStr` reference, but ran into lifetime errors
when trying to use a reference. Instead ended up making it a `CString`.

Still need to handle errors. My plan is to pass back an error string in
the result struct.

Remains to be seen whether this actually works when calling it over FFI,
but we have the beginnings of something.