Link library with a test C program
Make a test `includer.c` program that includes the Rust library and
calls our `c_run_key_action()` to see if it actually works. Currently it
doesn't, we get `(null)` printed to stdout.

Add a Makefile with the build command for the C program.

cbindgen.toml:
Remove `KeyActionResult` from exported types, as the `Option` field it
contains caused `gcc` to complain.

cocoa_bridge.rs:
* Comment out all 'cocoa' crate related code as the 'cocoa' code was
  interfering with GCC compilation as a result of my not linking to
  Cocoa frameworks.
* Add a new test map definition that corresponds with the one we look
  for in `includer.c`.

parser.rs:
Add `#[repr(C)]` to `MapKind` because it needs to be exported in the
library and generated into our C header file.