run_key_action_for_mode(): Start adding mode handling
Just the beginning of a way to get a map inside a mode. Not yet
complete. We need to modify `KeyActionResult` to give it an `in_mode`
field. This will enable us to tell the C code to use that mode scope.

Change the `mode` argument to `in_mode` because we already have a local
`mode` variable inside the function, and `in_mode` makes sense.