KeyActionResult: Make a pseudo-builder for the struct (WIP)
Make a builder for a new `KeyActionResult`, but without a new "builder"
type (seems like I should be able to get away without one in this simple
case).

I had make `parser::Trigger` public because at first I was using them in
`KeyActionResult` for the `in_mode` field. Oh yeah, there's a new
`in_mode` field on `KeyActionResult`. We'll be using that to store and
pass back the current mode to our FFI caller. This field can be used to
call `c_run_key_action()` with that mode scope.

Well, the idea is here, but I'm getting move errors:

    error[E0507]: cannot move out of borrowed content
      --> src/cocoa_bridge.rs:79:9
       |
    79 |         *self
       |         ^^^^^ cannot move out of borrowed content

    error[E0596]: cannot borrow immutable borrowed content as mutable
       --> src/cocoa_bridge.rs:162:30
        |
    162 |                               *KeyActionResult::new(MapKind::Map)
        |  ______________________________^
    163 | |                                 .with_action(&map.action)
        | |_________________________________________________________^ cannot borrow as mutable

    error[E0507]: cannot move out of borrowed content
       --> src/cocoa_bridge.rs:162:29
        |
    162 | /                             *KeyActionResult::new(MapKind::Map)
    163 | |                                 .with_action(&map.action)
    164 | |                                 .in_mode(trigger)
        | |_________________________________________________^ cannot move out of borrowed content

Looks like I'll need to change the `KeyActionResult` impl code to pass
by value instead of reference. Committing what I have now, though,
because it's been a while and it's become kind of a mess.