KeyActionResult: Fix reference/borrow errors
Return struct values instead of references in the `KeyActionResult` impl
to clear up the borrow errors from
a432dd2824499959635ac9a7cabec25a31dddb14.

Clean up some of the code. Remove the `build()` method because we no
longer need it. The rest of the methods will already return the
`KeyActionResult` we need.