Test mode finder
Add a new `moder.c` based on `includer.c` to test getting a mode and
then getting an action for a mode mapping.

Doesn't completely work currently. Right now all I get is this output:

    $ ./moder
    2
    Segmentation fault: 11

Clearly I'm messing up something with the pointer in
`c_run_key_action()`. Not sure what it is yet.

Changed `CKeyActionResult.in_mode` to be a `Trigger` instead of a
`HeadphoneButton` list. That allows us to pass it directly into the
second argument of `c_run_key_action()` without converting.

But maybe the pointer value isn't living long enough? Not sure what the
problem is yet.