run_key_action_for_mode(): Deactivate current mode using trigger
Enable mode deactivation. If there is a current mode, deactivate it if
the trigger pressed is the current mode's trigger.

Return a BS `KeyActionResult` just to satisfy the return type
requirements of the function. We should really clean this function up
and eliminate this return value, but we can't until we move map action
key simulator handling into Rust.