parser: Make a new `Action` type
With this new type, I'm trying to set up a way to parse actions more
precisely, to allow us to get special keys and modifier-flagged keys.

Still trying to work out how this is going to work at the parser level
though, since the action parser is going to need to depend on the map
kind parser.