Change `parser::Action` to an enum
It will now be a different type of `Vec` depending on whether it
represents a string map or a command. We'll have new parsers to parse
the action definition in a more fine-grained way.

The `String` variant is just there for temporary backward compatibility
while I figure out parsing for the other two.

Still need to update the tests for this change.