Get `action_map()` parser working
Additionally, fix a problem in `key_code()` when using special key names
that start with the same character (like any of the F keys, or Enter and
Esc). The parser would discard the first character, like in the
`definitions()` parser. Use the same solution here, by wrapping the
parser choices in `try()`s.

Add `Esc` to the test to confirm that it fails not just for the F keys
but for all key names that start similarly.