run_key_action_for_mode(): Run mode-level commands
Originally I was going to use `Action::Command` for this, which was
intended to hold a `Vec` of command arguments.

I decided against that approach. After thinking about command parsing a
bit, it's of course not just splitting into a `Vec` on whitespace. You
need to take into account quoting and whatever other idiosyncrasies.

Instead, I think I'm going to leave commands as `Action::String`s and
end up not using `Action::Command` for anything.

We take the shell command string from the action and feed it to a new
`Command`, executed using the `-c` option to the parent shell.

As I understand it, `spawn()` will reuse the parent process' stdout etc.
descriptors.