Remove return values from `c_run_key_action()` etc.
Get rid of the return values on `c_run_key_action()` and
`run_key_action_for_mode()`.

The reason why we returned a result previously was to feed information
back to the Objective-C caller to run subsequent parts of the pipeline,
like storing the current in-mode and simulating key presses. Now that
both of those are handled by the Rust code, we have no more need for
these return structs.

Get rid of the return values to clean up the code a bit.