Test media key simulator static library
Added a build of `libdome_key_event_source_simulator.a` to the root of
the project to test running the function from Rust. Works. Needs to be
built with custom `RUSTFLAGS` in order to build:

    $ RUSTFLAGS="-L ." cargo test dktest

We'll need to figure out how to add this to the Rust build script.

Now that we know it works, we can integrate the static library more
permanently into the project.