Add `NXKey` to `KeyboardKeyWithModifiers`
Add this new key type to the enum and enable it to be `tap()`ped like
the others.

Made `NXKey` a type alias instead of a struct to allow us to refer to
any of the NX key constants using the `NXKey` type. This also meant
moving those NX constants outside of the `NXKey` impl and into the top
level. May end up wrapping them back up in some kind of namespace later.

Adjusted the public-ness of some functions to support this change.