Config: Add `timeout` field
This new field will be used by the calling Objective-C code as the
timeout between multi-button maps.

It will be defined in a `config.toml` file in the XDG config directory,
with a default if unspecified. That part of the code hasn't been written
yet, but we've prepared by deriving Serde `Deserialize` on the `Config`
struct in order to deserialise a TOML config file.

The latest Serde is v1.0.80, but we can't use it because 'cbindgen'
depends on an older version. We had been using 'cbindgen' v0.6.3, which
depends on serde v1.0.21. Decided to update 'cbindgen' in order to use a
newer version of Serde, but that puts us at v1.0.58 instead of the
latest:

    error: failed to select a version for `serde_derive`.
        ... required by package `cbindgen v0.6.6`
        ... which is depended on by `dome-key-map v0.0.1 (file:///Users/tw/Documents/Development/dome-key-map)`
    versions that meet the requirements `= 1.0.58` are: 1.0.58

    all possible versions conflict with previously selected packages.

      previously selected package `serde_derive v1.0.80`
        ... which is depended on by `dome-key-map v0.0.1 (file:///Users/tw/Documents/Development/dome-key-map)`

    failed to select a version for `serde_derive` which could resolve this conflict