read_config_file(): Return a default config if no file found
The only config option we have is `timeout`, which has a default value.
Any future config options, it seems, should have default values also.

We shouldn't force users to create a config file, as the prior code
would do. They should be able to rely on the default, and create a
`config.toml` only if they want to change it.

If no config file is found, we should create a default `Config`.

Rename the method because now it still works without reading the config
file.