Add `trial` module, calculate days remaining
This new module will contain functions to calculate trial days.

Here we implement a function to calculate the days remaining in a trial
period.

If there are less than 0 days remaining, it returns an error result.

Couldn't use 'error-chain' because it was complaining about `Debug` and
`PartialEq` not being defined. Don't know if it's possible to implement
those traits on 'error-chain''s ErrorKind`, but it was complicated
enough that I ended up not bothering. Went with 'quick_error' instead to
create a custom error type without having to `impl` all the required
traits.