decode_datetime(): Return a `Result`
Make a custom error type because it's trickier to integrate with
'error-chain' given that 'magic-crypt''s `Error` type doesn't implement
the `std::error::Error` trait.

For the same reason, we can't use that as the `cause` for
`DateCryptError::Decrypt`.

This allows us to capture the errors without panicking.

Include some additional cleanups as a result of removing the `use
std::result` and `use errors::*` imports.