trial: Add a function to initialise a trial timestamp in a file
Save an encoded timestamp to a file. This function should be run once if
no trial file is found. The encoded timestamp can then be read on
subsequent launches to determine whether the trial period is still in
effect.

Also add a few function stubs for other actions that seem relevant.