trial: Implement `do_trial()`
This function serves as the entry point to all trial functionality. When
called, it will try to determine the number of days remaining in the
trial.

* If the trial period is in progress, the number of days remaining will
  be printed.
* If an error is encountered involving the trial file, or the trial has
  expired, an error message is printed and the program exits.
* Finally, if an unrelated error occurs, print the error without
  exiting.

Change `DateTime<Local>`s to `DateTime<FixedOffset>`s in order to have
uniform date types. Otherwise we run into trouble when comparing dates
and when querying functions with different date typs.

TODO: This fails the tests (unless your local timezone is UTC
presumably) because the local time gets stripped of its offset and the
correct offset gets reapplied, incorrectly shifting the time. Figure out
a way to have uniform datetime types without muffing up the timezone
offsets.