decode_datetime(): Fix incorrect encoded & IV string splitting
I was using `rsplitn` to split the encoded value into an encrypted
timestamp and initialisation vector. Originally, I had used `splitn`,
but upon discovering `rsplitn` decided to use that instead to ensure we
captured the initialisation vector first, and that anything else would
be captured to a single string.

Didn't realise, though, that `rsplitn` orders element in reverse order
compared to `splitn`. Correct the capture order.