Rename `cocoa_bridge.rs` to `ffi.rs`
When I first created the file, it was going to be used to call Cocoa
methods using the 'cocoa' crate. It's since turned into a module that
contains C functions called by Objective-C code. This new name makes
more sense.

We'll need to move one or two non-FFI functions outside of this module
for better organisation.