Test of playing an audio file
Success! Using the 'rodio' crate to play an audio file. Include the
audio file in the binary using `include_bytes!`. This makes it so we get
a single self-contained binary.

Struggled a bit with getting the reader/array to `Seek`, but finally
figured out `Cursor` and got it working. Cool.