ffi::run_key_action(): Add `PlayAudio` argument
Mirror the `map::run_key_action()` function and add an extra `PlayAudio`
argument, which will control whether audio cues are played when a mode
is activated or deactivated.

The Objective-C code will pass in this value based on the value of a
`Config` setting, which will have come from a command line argument.