parser::map_group(): Use default `MapGroup` as a base
This means that if any of `<Up>`, `<Play>`, or `<Down>` are undefined in
the mappings file definition, they will be set to their default action
values (as set in `MapGroup::default()`).

To get rid of the default, map the button trigger to `<Nop>`.

Update a test that didn't map `<Up>` to give it the default mapping.