dome_key_state_load_map_group(): Remove `expect`s
Don't on errors from reading or parsing the mappings file. Instead print
the error and parse an empty mappings string.

Took a lot of doing to get this working because of this error:

    error[E0597]: `dkmap` does not live long enough
      --> src/ffi.rs:72:44
       |
    72 |                     match MapGroup::parse(&dkmap)
       |                                            ^^^^^ borrowed value does not live long enough
    ...
    82 |                 },
       |                 - borrowed value only lives until here
       |
       = note: borrowed value must be valid for the static lifetime...

Finally got tired to trying things and just decided to store the
unparsed mappings string in `State` to get it to live long enough.