MapAction::parse(): Try to return a `Result` (WIP)
Want to be able to return a `Result` here so we can print the error from
the `ffi` module so we don't have to permit 'stderrlog' on this parser
module.

Beset with errors like this:

    error[E0506]: cannot assign to `self.action` because it is borrowed
       --> src/parser.rs:188:21
        |
    176 |             Action::String(ref s) => {
        |                            ----- borrow of `self.action` occurs here
    ...
    188 |                     self.action = action;
        |                     ^^^^^^^^^^^^^^^^^^^^ assignment to borrowed `self.action` occurs here

    error: aborting due to previous error

Even cloning the `Action` doesn't work because the Combine error
`Result` needs an `&str` reference to the parser input, and if we
reassign `self.action`, that string reference would disappear.

Need to figure out a different way of dealing with this, but I at least
want to commit what I have because the next step is going to be
something different.