parser: Add tests for error message
Check that the desired error messages appear for a couple extra invalid
mapping definitions.

Rename the existing `map_group_shows_error_in_middle_of_line()` test to
use the naming pattern of the two new tests.

The new tests check that the right errors appear in map actions, and for
missing the closing brace on a mode definition.